[CORE-1041] Convert many auth integration tests into unit tests
